fix(pkg): rename package e
