+++ internal/storage/db/postgresql/postgresql.go
@@ -8,7 +8,7 @@ import (
 	_ "github.com/jackc/pgx/stdlib" // use as driver for sqlx
 	"github.com/jmoiron/sqlx"
 
-	"github.com/Employee-s-file-cabinet/backend/pkg/e"
+	"github.com/Employee-s-file-cabinet/backend/pkg/errhelper"
 )
 
 const (
@@ -57,7 +57,7 @@ func NewDB(dsn string, opts ...Option) (*DB, error) {
 	}
 
 	if err != nil {
-		return nil, e.Wrap("new db", err)
+		return nil, errhelper.Wrap("new db", err)
 	}
 
 	db.SetMaxOpenConns(db.maxOpenConn)
